update main util

+++ scripts/util.py
@@ -1,27 +1,22 @@
 from pathlib import Path
 from typing import List
-
 from tqdm import tqdm
-
 from scripts.Subject import Subject
 
 
 def cp_aligned_frames(subjects: List[Subject], dst_dir: Path) -> None:
-    from scripts.workspace.WorkspaceEnum import WorkspaceStr
     import shutil
-
     hard_clean_dir(dst_dir)
-    for subject in tqdm(subjects, total=len(subjects), desc="copying subject`s aligned frames", unit="subject",
-                        miniters=1.0):
+    for subject in tqdm(subjects, total=len(subjects), desc="copying subject's aligned frames", unit="subject", miniters=1.0):
         for src_file in subject.aligned_frames().glob('*'):
-            tmp = dst_dir / (WorkspaceStr.subject.value + str(subject.id()) + "_" + src_file.stem + src_file.suffix)
+            tmp = dst_dir / f"{subject.root_dir().name}_{subject.id()}_{src_file.stem}{src_file.suffix}"
             shutil.copy(src_file, tmp)
 
 
 def hard_clean_dir(dst_dir: Path) -> None:
     import shutil
-
-    shutil.rmtree(dst_dir)
+    if dst_dir.exists():
+        shutil.rmtree(dst_dir)
     dst_dir.mkdir(exist_ok=True)
